Remove pi pulse train, fix qubit states angle
