+++ src/qibocal/calibrations/__init__.py
@@ -2,10 +2,10 @@ from qibocal.calibrations.characterization.allXY import *
 from qibocal.calibrations.characterization.calibrate_qubit_states import *
 from qibocal.calibrations.characterization.chevron import *
 from qibocal.calibrations.characterization.cryoscope import *
+from qibocal.calibrations.characterization.fast_reset import *
 from qibocal.calibrations.characterization.flipping import *
 from qibocal.calibrations.characterization.integration_weights_optimization import *
 from qibocal.calibrations.characterization.iqm_couplers import *
-from qibocal.calibrations.characterization.pi_pulse_train import *
 from qibocal.calibrations.characterization.qubit_spectroscopy import *
 from qibocal.calibrations.characterization.rabi import *
 from qibocal.calibrations.characterization.ramsey import *
