+++ src/qibocal/calibrations/characterization/flipping.py
@@ -14,8 +14,8 @@ def flipping(
     qubits: dict,
     nflips_max,
     nflips_step,
-    software_averages=1,
-    points=10,
+    nshots,
+    relaxation_time,
 ):
     r"""
     The flipping experiment correct the delta amplitude in the qubit drive pulse. We measure a qubit after applying
@@ -51,9 +51,6 @@ def flipping(
             - **qubit**: The qubit being tested
     """
 
-    # reload instrument settings from runcard
-    platform.reload_settings()
-
     # create a DataUnits object to store MSR, phase, i, q and the number of flips
     data = DataUnits(
         name="data",
@@ -62,64 +59,45 @@ def flipping(
     )
 
     # repeat the experiment as many times as defined by software_averages
-    count = 0
-    for iteration in range(software_averages):
-        # sweep the parameter
-        for flips in range(0, nflips_max, nflips_step):
-            # save data as often as defined by points
-            if count % points == 0 and count > 0:
-                # save data
-                yield data
-                # calculate and save fit
-                yield flipping_fit(
-                    data,
-                    x="flips[dimensionless]",
-                    y="MSR[uV]",
-                    qubits=qubits,
-                    resonator_type=platform.resonator_type,
-                    pi_pulse_amplitudes={
-                        q: qubits[qubit].pi_pulse_amplitude for q in qubits
-                    },
-                    labels=["amplitude_correction_factor", "corrected_amplitude"],
-                )
+    for flips in range(0, nflips_max, nflips_step):
+        # create a sequence of pulses for the experiment
+        sequence = PulseSequence()
+        ro_pulses = {}
+        for qubit in qubits:
+            RX90_pulse = platform.create_RX90_pulse(qubit, start=0)
+            sequence.add(RX90_pulse)
+            # execute sequence RX(pi/2) - [RX(pi) - RX(pi)] from 0...flips times - RO
+            start1 = RX90_pulse.duration
+            for j in range(flips):
+                RX_pulse1 = platform.create_RX_pulse(qubit, start=start1)
+                start2 = start1 + RX_pulse1.duration
+                RX_pulse2 = platform.create_RX_pulse(qubit, start=start2)
+                sequence.add(RX_pulse1)
+                sequence.add(RX_pulse2)
+                start1 = start2 + RX_pulse2.duration
 
-            # create a sequence of pulses for the experiment
-            sequence = PulseSequence()
-            ro_pulses = {}
-            for qubit in qubits:
-                RX90_pulse = platform.create_RX90_pulse(qubit, start=0)
-                sequence.add(RX90_pulse)
-                # execute sequence RX(pi/2) - [RX(pi) - RX(pi)] from 0...flips times - RO
-                start1 = RX90_pulse.duration
-                for j in range(flips):
-                    RX_pulse1 = platform.create_RX_pulse(qubit, start=start1)
-                    start2 = start1 + RX_pulse1.duration
-                    RX_pulse2 = platform.create_RX_pulse(qubit, start=start2)
-                    sequence.add(RX_pulse1)
-                    sequence.add(RX_pulse2)
-                    start1 = start2 + RX_pulse2.duration
+            # add ro pulse at the end of the sequence
+            ro_pulses[qubit] = platform.create_qubit_readout_pulse(qubit, start=start1)
+            sequence.add(ro_pulses[qubit])
 
-                # add ro pulse at the end of the sequence
-                ro_pulses[qubit] = platform.create_qubit_readout_pulse(
-                    qubit, start=start1
-                )
-                sequence.add(ro_pulses[qubit])
+        # execute the pulse sequence
+        results = platform.execute_pulse_sequence(
+            sequence,
+            nshots=nshots,
+            relaxation_time=relaxation_time,
+        )
 
-            # execute the pulse sequence
-            results = platform.execute_pulse_sequence(sequence)
-
-            for ro_pulse in ro_pulses.values():
-                # average msr, phase, i and q over the number of shots defined in the runcard
-                r = results[ro_pulse.serial].to_dict(average=True)
-                r.update(
-                    {
-                        "flips[dimensionless]": flips,
-                        "qubit": ro_pulse.qubit,
-                        "iteration": iteration,
-                    }
-                )
-                data.add(r)
-            count += 1
+        for ro_pulse in ro_pulses.values():
+            # average msr, phase, i and q over the number of shots defined in the runcard
+            r = results[ro_pulse.serial].to_dict(average=True)
+            r.update(
+                {
+                    "flips[dimensionless]": flips,
+                    "qubit": ro_pulse.qubit,
+                    "iteration": [0],
+                }
+            )
+            data.add(r)
     yield data
 
     yield flipping_fit(
