+++ src/qibocal/fitting/methods.py
@@ -1023,6 +1023,9 @@ def calibrate_qubit_states_fit(data, x, y, nshots, qubits, degree=True):
         iq_state1_rotated = iq_state1_translated * np.exp(-1j * rotation_angle)
         iq_state0_rotated = iq_state0_translated * np.exp(-1j * rotation_angle)
 
+        # iq_state1_rotated = iq_state1 * np.exp(-1j * rotation_angle)
+        # iq_state0_rotated = iq_state0 * np.exp(-1j * rotation_angle)
+
         real_values_state1 = iq_state1_rotated.real
         real_values_state0 = iq_state0_rotated.real
 
@@ -1052,7 +1055,7 @@ def calibrate_qubit_states_fit(data, x, y, nshots, qubits, degree=True):
         # assignment_fidelity = 1/2 + (cum_distribution_state1[argmax] - cum_distribution_state0[argmax])/nshots/2
 
         results = {
-            "rotation_angle": (-rotation_angle * 360 / (2 * np.pi)) % 360,  # in degrees
+            "rotation_angle": -rotation_angle,  # (-rotation_angle * 360 / (2 * np.pi)) % 360,  # in degrees
             "threshold": threshold,
             "fidelity": fidelity,
             "assignment_fidelity": assignment_fidelity,
