+++ src/qibocal/plots/__init__.py
@@ -2,6 +2,7 @@ from qibocal.plots.allXY import *
 from qibocal.plots.calibrate_qubit_states import *
 from qibocal.plots.chevron import *
 from qibocal.plots.cryoscope import *
+from qibocal.plots.fast_reset import *
 from qibocal.plots.flipping import *
 from qibocal.plots.rabi import *
 from qibocal.plots.ramsey import *
