+++ src/qibocal/plots/rabi.py
@@ -673,234 +673,3 @@ def duration_amplitude_msr_phase(folder, routine, qubit, format):
     figures.append(fig)
 
     return figures, fitting_report
-
-
-# Pi-pulse train oscillations
-def time_msr_phase_train(folder, routine, qubit, format):
-    figures = []
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    # iterate over multiple data folders
-    subfolders = get_data_subfolders(folder)
-    report_n = 0
-    fitting_report = ""
-    for subfolder in subfolders:
-        try:
-            data = DataUnits.load_data(folder, subfolder, routine, format, "data")
-            data.df = data.df[data.df["qubit"] == qubit]
-        except:
-            data = DataUnits(options=["qubit", "iteration", "n"])
-
-        try:
-            data_fit = Data.load_data(folder, subfolder, routine, format, f"fits")
-            data_fit.df = data_fit.df[data_fit.df["qubit"] == qubit]
-        except:
-            data_fit = Data(
-                quantities=[
-                    "popt0",
-                    "popt1",
-                    "popt2",
-                    "popt3",
-                    "popt4",
-                    "label1",
-                    "label2",
-                    "qubit",
-                ]
-            )
-
-        iterations = data.df["iteration"].unique()
-        ns = data.df["n"].unique()
-
-        if len(iterations) > 1:
-            opacity = 0.3
-        else:
-            opacity = 1
-
-        for iteration in iterations:
-            iteration_data = data.df[data.df["iteration"] == iteration]
-            fig.add_trace(
-                go.Scatter(
-                    x=iteration_data["n"],
-                    y=iteration_data["MSR"].pint.to("uV").pint.magnitude,
-                    marker_color=get_color(report_n),
-                    opacity=opacity,
-                    name=f"q{qubit}/r{report_n}",
-                    showlegend=not bool(iteration),
-                    legendgroup=f"q{qubit}/r{report_n}",
-                ),
-                row=1,
-                col=1,
-            )
-            fig.add_trace(
-                go.Scatter(
-                    x=iteration_data["n"],
-                    y=iteration_data["phase"].pint.to("rad").pint.magnitude,
-                    marker_color=get_color(report_n),
-                    opacity=opacity,
-                    name=f"q{qubit}/r{report_n}",
-                    showlegend=False,
-                    legendgroup=f"q{qubit}/r{report_n}",
-                ),
-                row=1,
-                col=2,
-            )
-        if len(iterations) > 1:
-            fig.add_trace(
-                go.Scatter(
-                    x=ns,
-                    y=data.df.groupby("time")["MSR"]
-                    .mean()
-                    .pint.to("uV")
-                    .pint.magnitude,
-                    marker_color=get_color(report_n),
-                    name=f"q{qubit}/r{report_n}: Average",
-                    showlegend=True,
-                    legendgroup=f"q{qubit}/r{report_n}: Average",
-                ),
-                row=1,
-                col=1,
-            )
-            fig.add_trace(
-                go.Scatter(
-                    x=ns,
-                    y=data.df.groupby("time")["phase"]  # pylint: disable=E1101
-                    .mean()
-                    .pint.to("rad")
-                    .pint.magnitude,
-                    marker_color=get_color(report_n),
-                    showlegend=False,
-                    legendgroup=f"q{qubit}/r{report_n}: Average",
-                ),
-                row=1,
-                col=2,
-            )
-
-        # add fitting trace
-        if len(data) > 0 and (qubit in data_fit.df["qubit"].values):
-            timerange = np.linspace(
-                min(data.get_values("time", "ns")),
-                max(data.get_values("time", "ns")),
-                2 * len(data),
-            )
-            params = data_fit.df[data_fit.df["qubit"] == qubit].to_dict(
-                orient="records"
-            )[0]
-            fig.add_trace(
-                go.Scatter(
-                    x=timerange,
-                    y=rabi(
-                        timerange,
-                        data_fit.get_values("popt0"),
-                        data_fit.get_values("popt1"),
-                        data_fit.get_values("popt2"),
-                        data_fit.get_values("popt3"),
-                        data_fit.get_values("popt4"),
-                    ),
-                    name=f"q{qubit}/r{report_n} Fit",
-                    line=go.scatter.Line(dash="dot"),
-                    marker_color=get_color(4 * report_n + 2),
-                ),
-                row=1,
-                col=1,
-            )
-
-            fitting_report = (
-                fitting_report
-                + (
-                    f"q{qubit}/r{report_n} | pi_pulse_duration: {params['pi_pulse_duration']:.2f} ns<br>"
-                )
-                + (
-                    f"q{qubit}/r{report_n} | pi_pulse_peak_voltage: {params['pi_pulse_peak_voltage']:,.0f} uV.<br><br>"
-                )
-            )
-
-        report_n += 1
-
-    # last part
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Number of RX",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Number of RX",
-        yaxis2_title="Phase (rad)",
-    )
-
-    figures.append(fig)
-
-    return figures, fitting_report
-
-
-# Fast reset oscillations
-def fast_reset_states(folder, routine, qubit, format):
-    figures = []
-
-    fig = make_subplots(
-        rows=1,
-        cols=1,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=("State",),
-    )
-
-    # iterate over multiple data folders
-    subfolders = get_data_subfolders(folder)
-    report_n = 0
-    fitting_report = ""
-    for subfolder in subfolders:
-        try:
-            data = DataUnits.load_data(folder, subfolder, routine, format, "data")
-            data.df = data.df[data.df["qubit"] == qubit]
-        except:
-            data = DataUnits(options=["qubit", "iteration", "n"])
-
-        iterations = data.df["iteration"]
-        states = data.df["MSR"].pint.to("V").pint.magnitude
-
-        opacity = 1
-
-        state0_count = states.value_counts()[0]
-        state1_count = states.value_counts()[1]
-        error = (state0_count - state1_count) / len(states)
-
-        fig.add_trace(
-            go.Scatter(
-                x=iterations,
-                y=states,
-                marker_color=get_color(report_n),
-                opacity=opacity,
-                name=f"q{qubit}/r{report_n}",
-                showlegend=True,
-                legendgroup=f"q{qubit}/r{report_n}",
-            ),
-            row=1,
-            col=1,
-        )
-
-        title_text = ""
-        title_text += f"q{qubit}/r{report_n} | Error: {error:.6f}<br>"
-
-        fitting_report = fitting_report + title_text
-        report_n += 1
-
-    # last part
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Shot",
-        yaxis_title="State",
-    )
-
-    figures.append(fig)
-
-    return figures, fitting_report
